Move common code to CommonTestUtils
