+++ source/Octopus.Tentacle.CommonTestUtils/Diagnostics/InMemoryLog.cs
@@ -9,7 +9,7 @@ using NUnit.Framework;
 using Octopus.Diagnostics;
 using Octopus.Tentacle.Diagnostics;
 
-namespace Octopus.Tentacle.Tests.Integration.Support
+namespace Octopus.Tentacle.CommonTestUtils.Diagnostics
 {
     public class InMemoryLog : SystemLog
     {
