+++ source/Octopus.Tentacle.CommonTestUtils/Octopus.Tentacle.CommonTestUtils.csproj
@@ -24,6 +24,12 @@
     </Choose>
     <ItemGroup>
       <ProjectReference Include="..\Octopus.Tentacle.Contracts\Octopus.Tentacle.Contracts.csproj" />
+      <ProjectReference Include="..\Octopus.Tentacle\Octopus.Tentacle.csproj" />
+    </ItemGroup>
+    <ItemGroup>
+      <PackageReference Include="FluentAssertions" Version="6.7.0" />
+      <PackageReference Include="NUnit" Version="3.13.3" />
+      <PackageReference Include="Serilog" Version="2.12.0" />
     </ItemGroup>
 
 </Project>
