+++ source/Octopus.Tentacle.Tests.Integration/Startup/WindowsServiceConfiguratorFixture.cs
@@ -8,8 +8,8 @@ using System.ServiceProcess;
 using FluentAssertions;
 using NSubstitute;
 using NUnit.Framework;
+using Octopus.Tentacle.CommonTestUtils.Diagnostics;
 using Octopus.Tentacle.Startup;
-using Octopus.Tentacle.Tests.Integration.Support;
 using Octopus.Tentacle.Tests.Integration.Support.TestAttributes;
 using Octopus.Tentacle.Tests.Integration.Util;
 using Octopus.Tentacle.Util;
