+++ source/Octopus.Tentacle.Tests.Integration/Support/ClientAndTentacle.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Halibut;
 using Octopus.Tentacle.Client;
 using Octopus.Tentacle.Client.Retries;
+using Octopus.Tentacle.CommonTestUtils;
 using Octopus.Tentacle.Tests.Integration.Support.Legacy;
 using Octopus.TestPortForwarder;
 using Serilog;
