+++ source/Octopus.Tentacle.Tests.Integration/Support/RunningTentacle.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Octopus.Tentacle.Client.Retries;
+using Octopus.Tentacle.CommonTestUtils;
 using Polly;
 using Serilog;
